renamings, typos
