+++ abstract_domains/numerical/numerical.py
@@ -3,13 +3,16 @@ from core.expressions import VariableIdentifier, Expression
 
 
 class NumericalMixin(metaclass=ABCMeta):
+    """A mixin to add a numeric interface to abstract domains.
+    """
+
     @abstractmethod
     def forget(self, var: VariableIdentifier):
-        """Forget all constraints about a variable."""
+        """Forget all information about a variable."""
 
     @abstractmethod
     def set_bounds(self, var: VariableIdentifier, lower: int, upper: int):
-        """Sets the upper and lower bound of a variable.
+        """Set the upper and lower bound of a variable.
         
         :param var: the variable of interest
         :param lower: the upper bound of ``var``
@@ -17,11 +20,11 @@ class NumericalMixin(metaclass=ABCMeta):
 
     @abstractmethod
     def get_bounds(self, var: VariableIdentifier):
-        """Returns the upper and lower bound of a variable.
+        """Return the upper and lower bound of a variable.
         
         :param var: the variable of interest
         :return: A tuple (lower, upper) of the bounds of ``var``"""
 
     @abstractmethod
     def evaluate(self, expr: Expression):
-        """Evaluates the expression within this abstract domain."""
+        """Evaluate the expression within this numerical domain."""
